snap-bootstrap: update rename
